clauditor-kut.9: Split auth-related test classes into tests/test_providers_auth.py (#144 US-009)
Move auth-helper test classes out of tests/test_providers_anthropic.py
into a new tests/test_providers_auth.py, mirroring the production-code
split between _providers/_anthropic.py (SDK seam) and _providers/_auth.py
(auth helpers + announcement family).

Moved classes:
- TestCheckAnyAuthAvailable
- TestCheckApiKeyOnly
- TestAnnounceImplicitNoApiKey
- TestCallAnthropicDeprecationAnnouncement
- TestExceptionClassIdentity

Also moved the _patch_which helper which is only used by the moved
classes. The CLI-transport TestStderrAnnouncement class stays in the
SDK-seam file (it tests _announced_cli_transport, not auth coupling).

Imports in each file are scoped — orphan auth-helper imports removed
from test_providers_anthropic.py, and the new file imports only the
auth symbols its classes use.

All 139 tests in both files pass; full suite passes with 98.45%
coverage.
